Move TypeScript dependencies
Move TypeScript dependencies to devDependencies for optimized the production build

To keep the production build lean and better organized, `typescript` and `ts-node` dependencies were moved from `dependencies` to `devDependencies` in `package.json`. These tools are only required during development and the build phase and are not necessary for the production runtime.

Changes:
- Moved `typescript` and `ts-node` from `dependencies` to `devDependencies` in `package.json`.